Support ScorePoseTask instantiation in io namespace
